+++ user-management/backend/app/user-gateway/middleware/auth_middleware.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/gofiber/fiber/v2"
-	"tmossDev.github.com/eco-system/shared-components/backend/domain/user/constants"
+	"tmossDev.github.com/eco-system/shared-components/backend/package/user/constants"
 	"tmossDev.github.com/eco-system/shared-components/backend/package/utils"
 	"tmossDev.github.com/eco-system/user-management/backend/domain/user/service"
 )